data updates based on playing with Yaesu FT5DR.
I couldn't find any specific documentation, so these updates are largely
just from playing around with a Yaesu FT5DR. I added a few channels to
memory manually on the radio, then saved to a microSD card. Then I
started using this tool to parse the MEMORY.dat that was saved to the
microSD card, and export the parsed data to test.csv and test.dat files.
Using xxd and diff, I kept making updates until the data was as similar
as possible.