csv: change format to match Repeater Book.
Update the CSV format to match Repeater Book's Yaesu FT3 ADMS 11
export format. It seems this format might have been added to Repeater
Book after this tool was developed, but I'm not exactly sure on that.
Anyhow, the goal is essentially to be able to export from Repeater Book
and then use this tool to create a MEMORY.dat that can be loaded onto
the radio via microSD. There are more fields and they are in a different
order than the format this tool was using, but one of the benefits is
that we can capture more of the data fields so that we can import and
export without too much being changed the process.

Repeater Book documents their Yaesu FT3 ADMS 11 CSV fields at
https://www.repeaterbook.com/wiki/doku.php?id=yaesu_adms_11.